requirements.txt
# Core dependencies
flask==2.3.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
redis==4.6.0
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0

# AI and ML
openai==1.3.0
google-generativeai==0.3.2
autogen-ext==0.0.1
autogen-agentchat==0.0.1

# Web scraping
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.1
numpy==1.24.3
python-dateutil==2.8.2

# Notifications
slack-sdk==3.21.3
twilio==8.10.0

# Development and testing
pytest==7.4.2
pytest-asyncio==0.21.1

